chore: improve ApplePay module redability
