+++ ios/Components/ApplePay/ApplePayModuleUtilities.swift
@@ -1,0 +1,64 @@
+import Adyen
+import PassKit
+
+extension ApplePayPaymentMethod {
+
+    internal var supportedNetworks: [PKPaymentNetwork] {
+        var networks = PKPaymentRequest.availableNetworks()
+        if let brands {
+            let brandsSet = Set(brands)
+            networks = networks.filter { brandsSet.contains($0.txVariantName) }
+        }
+        return networks
+    }
+
+}
+
+extension PKPaymentNetwork {
+
+    internal var txVariantName: String {
+        switch self {
+            case .masterCard: return "mc"
+            case .cartesBancaires: return "cartebancaire"
+            default: return self.rawValue.lowercased()
+        }
+    }
+
+}
+
+extension PKContact {
+    var jsonObject: [String: Any] {
+        var dictionary: [String: Any] = [:]
+
+        if let email = self.emailAddress {
+            dictionary[ApplePayKeys.Contact.emailAddress] = email
+        }
+
+        if let phoneNumber = self.phoneNumber {
+            dictionary[ApplePayKeys.Contact.phoneNumber] = phoneNumber.stringValue
+        }
+
+        if let name = self.name {
+            dictionary[ApplePayKeys.Contact.givenName] = name.givenName
+            dictionary[ApplePayKeys.Contact.familyName] = name.familyName
+        }
+
+        if let name = self.name?.phoneticRepresentation {
+            dictionary[ApplePayKeys.Contact.phoneticGivenName] = name.givenName
+            dictionary[ApplePayKeys.Contact.phoneticFamilyName] = name.familyName
+        }
+
+        if let postalAddress = self.postalAddress {
+            dictionary[ApplePayKeys.Contact.addressLines] = postalAddress.street
+            dictionary[ApplePayKeys.Contact.subLocality] = postalAddress.subLocality
+            dictionary[ApplePayKeys.Contact.locality] = postalAddress.city
+            dictionary[ApplePayKeys.Contact.postalCode] = postalAddress.postalCode
+            dictionary[ApplePayKeys.Contact.subAdministrativeArea] = postalAddress.subAdministrativeArea
+            dictionary[ApplePayKeys.Contact.administrativeArea] = postalAddress.state
+            dictionary[ApplePayKeys.Contact.country] = postalAddress.country
+            dictionary[ApplePayKeys.Contact.countryCode] = postalAddress.isoCountryCode
+        }
+
+        return dictionary
+    }
+}
